lesson_5/pyproject.toml
[project]
name = "lesson-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.2.1",
]
